backend/requirements.txt
Django==5.0
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
django-q2==1.6.1
croniter==2.0.1

boto3==1.34.11
redis==5.0.1

psycopg2-binary==2.9.11
sentry-sdk==1.39.2
requests==2.31.0

torch
numpy
scipy
protobuf
accelerate
sentencepiece
transformers
diffusers
Pillow===10.4.0
pygltflib